added autostart.sh
Pretty much a muk of what I use for starting up my session. Since my monitor setup is super weird, this really helps. Dropbox is no longer installed (or supported) so its pointless to keep it here but eh..Also Screencloud appears to autostart even without this but there doesnt appear to be any conflicts so I guess it's fine? Who knows. This is mostly helpful for nitrogen, tint2, and monitor setup. You're more than welcome to use this to help you out or whatever. Whoosh!