added menu.xml
Semi-first backup of this thing here. Damn I love it! However because this cannot be easily carried over to each new system I install or environment I dump this on, this isnt very useful. Well it is, but its not. Kinda like it's a nice template of how I like stuff, but it's kinda dumb to continue to copy/paste something like this and re-edit it to fit each system. Either way a backup is a backup. Not like I can't backup future edits/versions for each system/station!